api/app/pyproject.toml
[tool.poetry]
name = "mythica-app"
version = "1.0.0"
description = "FastAPI service to provide asset services."
authors = ["Jacob Repp <jacobrepp@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/MythicaAI/infra/api/app"
repository = "https://github.com/MythicaAI/infra"
keywords = ["game assets", "file management", "sqlite", "cache", "graph", "games"]
package-mode = false

[tool.poetry.dependencies]
python = "<4.0,>=3.11"
minio = "^7.2.15"
sqlalchemy = "2.0.38"
asyncpg = "^0.30.0"
google-auth = "^2.38.0"
google-cloud-storage = "^3.0.0"
fastapi = "^0.115.8"
pydantic-settings = "^2.3.0"
python-multipart = "0.0.19"
jinja2 = "^3.1.5"
sqlmodel = "0.0.22"
prometheus-fastapi-instrumentator = "^7.0.0"
typing-extensions = "^4.12.2"
alembic = "^1.14.1"
asyncio = "^3.4.3"
python-json-logger = "2.0.7"
base58 = "^2.1.1"
authlib = "^1.4.1"
pyjwt = { version = "2.9.0", extras = ["crypto"] }
fastapi-auth0 = "^0.5.0"
starlette = "0.41.2"
tzdata = "2024.2"
websockets = "13.1"
redis = "5.2.0"
ripple = { path = "../../libs/python/ripple", develop = true }
cryptid = { path = "../../libs/python/cryptid", develop = true }
better-profanity = "^0.7.0"
opentelemetry-instrumentation = "0.51b0"
opentelemetry-instrumentation-fastapi = "0.51b0"
opentelemetry-instrumentation-httpx = "0.51b0"
opentelemetry-exporter-otlp = "^1.27.0"
opentelemetry-distro = "0.51b0"
opentelemetry-api = "^1.27.0"
opentelemetry-sdk = "^1.27.0"
opentelemetry-instrumentation-sqlalchemy = "0.51b0"
uvicorn = "^0.34.0"
pydantic = "2.9.2"
email-validator = "^2.2.0"
opentelemetry-exporter-gcp-trace = "^1.8.0"
sqltap-ng = "^0.3.11"
greenlet = "^3.1.1"
aiohappyeyeballs = "^2.4.6"
aiosignal = "^1.3.2"
annotated-types = "^0.7.0"
anyio = "^4.8.0"
argon2-cffi = "^23.1.0"
argon2-cffi-bindings = "^21.2.0"
asgiref = "^3.8.1"
attrs = "^25.1.0"
cachetools = "^5.5.1"
certifi = "^2025.1.31"
cffi = "^1.17.1"
charset-normalizer = "^3.4.1"
pytest-asyncio = "^0.25.3"


[tool.poetry.extras]
sqlalchemy = ["asyncio"]
authlib = []

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
invoke = "^2.2.0"
flake8 = "^7.1.1"
httpx = "^0.28.1"
munch = "^4.0.0"
pylint = "^3.3.4"
pytest-cov = "^6.0.0"
pytest-timeout = "^2.3.1"
httpx-ws = "^0.7.1"
aiosqlite = "^0.21.0"

[tool.poetry.scripts]
start = "main:start"

[build-system]
requires = ["poetry-core>=1.8.5"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.report]
omit = [
    "tests/*",
]

[tool.pytest.ini_options]
addopts = "-p no:logging"
asyncio_mode = "auto"
